refactor: move order entity to entity packages
