+++ apps/seller/lib/presentation/screen/orders/cubit/seller_orders_state.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 
-import '../../../../domain/entities/seller_orders_filter.dart';
 import '../../../../domain/entity/seller_order.dart';
+import '../../../../domain/entity/seller_orders_filter.dart';
 
 sealed class SellerOrdersState extends Equatable {
   const SellerOrdersState();
